+++ src/main/java/com/ml/clima/utils/CalculadorPosicion.java
@@ -7,9 +7,9 @@ import java.util.HashMap;
 
 public final class CalculadorPosicion {
 
-    public static HashMap<String, Float[]> calcularPosicion(Integer dia){
+    public static HashMap<String, Double[]> calcularPosicion(Integer dia){
         //TODO implementar logica
-        HashMap<String,Float[]> map = new HashMap<String, Float[]>();
+        HashMap<String,Double[]> map = new HashMap<String, Double[]>();
         map.put(PlanetaEnum.FARENGI.name(),CalculadorFactory.getPosicion(PlanetaEnum.FARENGI.name(),dia));
         map.put(PlanetaEnum.VULCANO.name(),CalculadorFactory.getPosicion(PlanetaEnum.VULCANO.name(),dia));
         map.put(PlanetaEnum.BETASOIDE.name(),CalculadorFactory.getPosicion(PlanetaEnum.BETASOIDE.name(),dia));
